Update 4.0 Final Update
In this update Robert and I tried to optimise the game as much as we could in the given time space. We also added a new colouring window in which the player can set the colour of the snake before playing it. We also added the links that we used during writing the code. The game is done.

Is it?